Entity id 컬럼 수정(ex class_id -> id) 변경에 따른 코드 수정

+++ src/main/kotlin/reservation/project/domain/academy/entity/Academy.kt
@@ -9,7 +9,7 @@ data class Academy(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    val academyId: Int = 0,
+    val academyId: Long = 0,
 
     @Column(name = "academy_name", nullable = false)
     val academyName: String,
